requirements.txt
gradio
python-dotenv
openai
instructor
jinja2
pydantic
